Delete Comment For Helper Function
Got rid of the comment for valuePairExistsInAttributePair helper function. Might want to write the comment for the function at another point since it does the check kind of weirdly. It's at least different to some of the other helper functions Greg has seen in the examples repo. Knowing what it's used for would be good.